# lplates/openapi.yaml
openapi: 3.0.3
info:
  title: lplates API
  description: >
    REST API for the lplates driving-lesson booking platform. Covers user
    authentication, instructor management, lesson bookings, EDT (Essential
    Driver Training) progress tracking, instructor pricing, and feedback.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /instructors/:
    get:
      tags: [Instructors]
      summary: Get all instructors
      responses:
        '200':
          description: List of instructors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instructor'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /instructors/{instructorId}:
    get:
      tags: [Instructors]
      summary: Get a single instructor by instructor ID
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: Instructor found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /instructors/create:
    post:
      tags: [Instructors]
      summary: Create an instructor profile (INSTRUCTOR, ADMIN)
      description: The `userId` field must match the authenticated user's ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Instructor'
      responses:
        '200':
          description: Instructor created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Instructor created successfully with ID: 4"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          description: Failed to create instructor
  /instructors/{instructorId}/lessons:
    get:
      tags: [Instructors]
      summary: Get an instructor with their full lesson list embedded
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: Instructor with lessons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /instructors/pending:
    get:
      tags: [Instructors]
      summary: Get all instructors awaiting approval (ADMIN only)
      responses:
        '200':
          description: List of pending instructors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instructor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /instructors/{instructorId}/approval:
    patch:
      tags: [Instructors]
      summary: Approve or reject an instructor sign-up (ADMIN only)
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalStatusRequest'
      responses:
        '200':
          description: Updated instructor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '400':
          description: Invalid approvalStatus value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /instructors/{instructorId}/profile-picture:
    post:
      tags: [Instructors]
      summary: Upload an instructor profile picture (owning INSTRUCTOR or ADMIN)
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG, or WEBP image, max 5MB
              required: [file]
      responses:
        '200':
          description: Updated instructor with new profilePicture URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '400':
          description: Empty file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the owning instructor and not an ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /lessons/instructor/{instructorId}:
    get:
      tags: [Lessons]
      summary: Get all lessons for an instructor
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '204':
          description: No lessons found
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /pricing/create:
    post:
      tags: [Pricing]
      summary: Create an instructor pricing entry (INSTRUCTOR, ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorPricing'
      responses:
        '200':
          description: Created pricing entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorPricing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /pricing/{pricingId}:
    get:
      tags: [Pricing]
      summary: Get a pricing entry by ID
      parameters:
        - name: pricingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pricing entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorPricing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pricing]
      summary: Delete a pricing entry
      parameters:
        - name: pricingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /pricing/instructor/{instructorId}:
    get:
      tags: [Pricing]
      summary: Get all pricing entries for an instructor
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: List of pricing entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstructorPricing'
        '204':
          description: No pricing entries found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /feedback/create:
    post:
      tags: [Feedback]
      summary: Create feedback for a lesson
      description: One feedback per lesson (unique constraint). `rating` is 1-5.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feedback'
      responses:
        '200':
          description: Created feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /feedback/{feedbackId}:
    get:
      tags: [Feedback]
      summary: Get feedback by ID
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Feedback found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /feedback/lesson/{lessonId}:
    get:
      tags: [Feedback]
      summary: Get feedback for a lesson
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Feedback found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /feedback/instructor/{instructorId}:
    get:
      tags: [Feedback]
      summary: Get all feedback for an instructor's lessons
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
        '204':
          description: No feedback found
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    InstructorId:
      name: instructorId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    LessonId:
      name: lessonId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid username or password"
...
    LessonType:
      type: string
      enum: [lesson, edt, test_car_hire]
    ApprovalStatus:
      type: string
      enum: [pending, approved, rejected]
...
    Instructor:
      type: object
      properties:
        instructorId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        approvalStatus:
          $ref: '#/components/schemas/ApprovalStatus'
        gender:
          type: string
        phoneNumber:
          type: string
        adiNumber:
          type: string
        transmission:
          type: string
          example: manual
        yearsExperience:
          type: integer
        rating:
          type: number
          format: double
        carMake:
          type: string
        carModel:
          type: string
        locations:
          type: array
          items:
            type: string
        profilePicture:
          type: string
          nullable: true
          description: URL relative to the API base, e.g. /uploads/instructors/1/abc.jpg
        description:
          type: string
        reviewsCount:
          type: integer
        agreeTerms:
          type: boolean
        offersTestCarHire:
          type: boolean
        testCarHirePrice:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'
    ApprovalStatusRequest:
      type: object
      required: [approvalStatus]
      properties:
        approvalStatus:
          $ref: '#/components/schemas/ApprovalStatus'
    Lesson:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        scheduledDate:
          type: string
          format: date
          example: "2026-08-10"
        scheduledTime:
          type: string
          example: "09:00:00"
        durationMinutes:
          type: integer
          example: 60
        status:
          type: string
          example: pending
        paymentStatus:
          type: string
          example: unpaid
        price:
          type: number
          format: double
          example: 60.00
        currency:
          type: string
          example: EUR
        lessonType:
          $ref: '#/components/schemas/LessonType'
        edtModule:
          type: string
        edtCompleted:
          type: boolean
        notes:
          type: string
        cancellationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InstructorPricing:
      type: object
      properties:
        id:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        durationMinutes:
          type: integer
          example: 60
        price:
          type: number
          format: double
          example: 60.00
    Feedback:
      type: object
      properties:
        id:
          type: integer
          format: int64
        lessonId:
          type: integer
          format: int64
        authorId:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
...